+++ inmoose/utils/__init__.py
@@ -1,3 +1,4 @@
 from .factor import Factor, gl, asfactor
 from .lm import lm_fit, lm_wfit
+from .splines import ns, spline_design
 from .stats import rnbinom, dnbinom_mu, dnorm, pnorm, pt